make DataSource cache internal
