+++ core/src/tile/tileID.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <functional> // for hash function
 #include <string>
 
 /* An immutable identifier for a map tile 
@@ -69,25 +68,4 @@ struct TileID {
 
 static TileID NOT_A_TILE(-1, -1, -1);
 
-}
-
-// The generic hash_combine used in Boost
-// http://www.boost.org/doc/libs/1_35_0/doc/html/boost/hash_combine_id241013.html
-template <class T>
-inline void hash_combine(std::size_t & seed, const T & v) {
-    std::hash<T> hasher;
-    seed ^= hasher(v) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
-}
-
-namespace std {
-    template <>
-    struct hash<Tangram::TileID> {
-        size_t operator()(const Tangram::TileID& k) const {
-            std::size_t seed = 0;
-            hash_combine(seed, k.x);
-            hash_combine(seed, k.y);
-            hash_combine(seed, k.z);
-            return seed;
-        }
-    };
 }
